refactor: Update import paths for generated API client models
