--- practice/module2/lesson3/rick-and-morty-api/src/api/generate-client.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-openapi-generator generate -i ./spec.yml -g typescript-fetch -o ./api-client-generated
