+++ practice/module2/lesson3/rick-and-morty-api/src/api/models/CharacterLocation.ts
