--- practice/module2/lesson3/rick-and-morty-api/src/api/validate-spec.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-openapi-generator validate -i ./spec.yml
